Rename ListView files, remove unused CSS styles
